IRENO Smart Assistant/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchain-core==0.2.43
langchain-text-splitters==0.2.4
openai==1.54.4
python-dotenv==1.0.0
requests==2.32.5
httpx==0.24.1
azure-storage-blob==12.19.0
